+++ GameGuy/include/Application/AudioManager.h
@@ -49,7 +49,7 @@ public:
 		if (d != devices.end())
 		{
 			// Device is available
-			int nDeviceID = distance(devices.begin(), d);
+			size_t nDeviceID = distance(devices.begin(), d);
 			WAVEFORMATEX waveFormat;
 			waveFormat.wFormatTag = WAVE_FORMAT_PCM;
 			waveFormat.nSamplesPerSec = m_nSampleRate;
@@ -60,7 +60,7 @@ public:
 			waveFormat.cbSize = 0;
 
 			// Open Device if valid
-			if (waveOutOpen(&m_hwDevice, nDeviceID, &waveFormat, (DWORD_PTR)waveOutProcWrap, (DWORD_PTR)this, CALLBACK_FUNCTION) != S_OK)
+			if (waveOutOpen(&m_hwDevice, (UINT)nDeviceID, &waveFormat, (DWORD_PTR)waveOutProcWrap, (DWORD_PTR)this, CALLBACK_FUNCTION) != S_OK)
 				return Destroy();
 		}
